+++ react-demo-app/src/components/Dashboard.tsx
@@ -2,6 +2,7 @@ import './Dashboard.css';
 import Softphone from './Softphone';
 import Devices from './Devices';
 import Header from './Header';
+import NoiseTester from './NoiseTester';
 
 
 export default function Dashboard() {
@@ -10,6 +11,7 @@ export default function Dashboard() {
       <Header></Header>
       <Devices></Devices>
       <Softphone></Softphone>
+      <NoiseTester></NoiseTester>
     </div>
   )
 }